reorg tests pass
