+++ commands/command_prs.go
@@ -4,6 +4,8 @@ import (
 	"flag"
 	"io"
 	"log/slog"
+
+	ex "github.com/joshallenit/stacked-diff/execute"
 )
 
 func createPrsCommand(stdOut io.Writer) Command {
